vercel changes

+++ database/migrations/2026_02_09_050133_create_health_scores_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('health_scores', function (Blueprint $table) {
             $table->id();
             $table->foreignId('store_id')->constrained()->cascadeOnDelete();
-            $table->integer('score'); // 0-100
+            $table->decimal('score', 5, 2); // 0.00-100.00
             $table->float('metric_availability');
             $table->float('metric_performance');
             $table->float('metric_incidents');
